+++ CHANGELOG.md
@@ -2,9 +2,15 @@
 
 ## v1.5.0
 
+### New Features
+
 - Adds a one-time execution option (`once`) to `EventListener` [#32](https://github.com/TENSIILE/saborter/pull/32)
 - Adds an extended stack for the `AbortError` error [#33](https://github.com/TENSIILE/saborter/pull/33)
-- Adds a `reason` option to `TimeoutError`
+- Adds a `reason` option to `TimeoutError` [#34](https://github.com/TENSIILE/saborter/pull/34)
+
+### Bug Fixes
+
+- Fixes a bug where the request state would not change to `cancelled` [#35](https://github.com/TENSIILE/saborter/pull/35)
 
 ## v1.4.2 (January 22th, 2026)